requirements.txt
streamlit==1.29.0
diffusers==0.24.0
transformers==4.36.0
torch==2.1.0
torchvision==0.16.0
accelerate==0.25.0
safetensors==0.4.0
pillow==10.1.0
numpy==1.24.3
matplotlib==3.7.2
opencv-python==4.8.1.78
scikit-image==0.21.0
